FEAT: transformStringToNumber 함수 share util 로 변경
